requirements.txt
numpy==1.24.3
pandas==2.0.3
yfinance==0.2.28
ta-lib==0.4.28
scikit-learn==1.3.0
matplotlib==3.7.2
pytest==7.4.0
python-dotenv==1.0.0
ccxt==4.0.0
pandas-ta==0.3.14b
statsmodels==0.14.0
scipy==1.11.3
optuna==3.3.0
backtrader==1.9.78.123
pymongo==4.5.0
redis==5.0.1
dash==2.13.0
plotly==5.17.0